tidy up grouping command
