+++ src/main/java/Main.java
@@ -4,7 +4,6 @@ import com.jagrosh.jdautilities.command.CommandClient;
 import com.jagrosh.jdautilities.command.CommandClientBuilder;
 import com.jagrosh.jdautilities.commons.waiter.EventWaiter;
 import command.*;
-import command.UserPlaylistCommand.CreatePlaylistCommand;
 import command.general.CustomPrefixCommand;
 import command.general.HelpCommand;
 import command.general.InviteCommand;
@@ -84,29 +83,29 @@ public class Main {
         commandClientBuilder.addCommand(new ShuffleCommand(nano));
         commandClientBuilder.addCommand(new RecommendationCommand(nano, youtubeClient));
         commandClientBuilder.addCommand(new LyricCommand(nano));
+        commandClientBuilder.addCommand(new SaveQueueToPlaylistCommand());
         commandClientBuilder.addCommand(new PremiumUserCommand());
         commandClientBuilder.addCommand(new UserHistoryCommand());
-        commandClientBuilder.addCommand(new CreatePlaylistCommand());
-        commandClientBuilder.addCommand(new command.UserPlaylistCommand.RenamePlaylistCommand());
-        commandClientBuilder.addCommand(new command.UserPlaylistCommand.DeletePlaylistCommand());
-        commandClientBuilder.addCommand(new command.UserPlaylistCommand.ShowPlaylistCommand());
-        commandClientBuilder.addCommand(new command.UserPlaylistCommand.AddTrackToPlaylistCommand(youtubeClient));
-        commandClientBuilder.addCommand(new command.UserPlaylistCommand.DeleteTrackFromPlaylistCommand());
-        commandClientBuilder.addCommand(new command.UserPlaylistCommand.ShowPlaylistTrackCommand());
-        commandClientBuilder.addCommand(new command.UserPlaylistCommand.PlayPlaylistCommand(nano));
-        commandClientBuilder.addCommand(new command.UserPlaylistCommand.SaveQueueToPlaylist());
+        commandClientBuilder.addCommand(new command.playlist.user.CreatePlaylistCommand());
+        commandClientBuilder.addCommand(new command.playlist.user.RenamePlaylistCommand());
+        commandClientBuilder.addCommand(new command.playlist.user.DeletePlaylistCommand());
+        commandClientBuilder.addCommand(new command.playlist.user.ShowPlaylistCommand());
+        commandClientBuilder.addCommand(new command.playlist.user.AddTrackToPlaylistCommand(youtubeClient));
+        commandClientBuilder.addCommand(new command.playlist.user.DeleteTrackFromPlaylistCommand());
+        commandClientBuilder.addCommand(new command.playlist.user.ShowPlaylistTrackCommand());
+        commandClientBuilder.addCommand(new command.playlist.user.PlayPlaylistCommand(nano));
 
         // Premium Guild Commands
 //        commandClientBuilder.addCommand(new PremiumGuildCommand());
 //        commandClientBuilder.addCommand(new GuildHistoryCommand(nano));
-//        commandClientBuilder.addCommand(new command.GuildPlaylistCommand.CreatePlaylistCommand());
-//        commandClientBuilder.addCommand(new command.GuildPlaylistCommand.RenamePlaylistCommand());
-//        commandClientBuilder.addCommand(new command.GuildPlaylistCommand.DeletePlaylistCommand());
-//        commandClientBuilder.addCommand(new command.GuildPlaylistCommand.ShowPlaylistCommand());
-//        commandClientBuilder.addCommand(new command.GuildPlaylistCommand.AddTrackToPlaylistCommand());
-//        commandClientBuilder.addCommand(new command.GuildPlaylistCommand.DeleteTrackFromPlaylistCommand());
-//        commandClientBuilder.addCommand(new command.GuildPlaylistCommand.ShowPlaylistTrackCommand());
-//        commandClientBuilder.addCommand(new command.GuildPlaylistCommand.PlayPlaylistCommand(nano));
+//        commandClientBuilder.addCommand(new command.playlist.guild.CreatePlaylistCommand());
+//        commandClientBuilder.addCommand(new command.playlist.guild.RenamePlaylistCommand());
+//        commandClientBuilder.addCommand(new command.playlist.guild.DeletePlaylistCommand());
+//        commandClientBuilder.addCommand(new command.playlist.guild.ShowPlaylistCommand());
+//        commandClientBuilder.addCommand(new command.playlist.guild.AddTrackToPlaylistCommand());
+//        commandClientBuilder.addCommand(new command.playlist.guild.DeleteTrackFromPlaylistCommand());
+//        commandClientBuilder.addCommand(new command.playlist.guild.ShowPlaylistTrackCommand());
+//        commandClientBuilder.addCommand(new command.playlist.guild.PlayPlaylistCommand(nano));
 
         // Owner's commands
         commandClientBuilder.addCommand(new ChangePresenceCommand());
