+++ src/main/java/command/playlist/guild/AddTrackToPlaylistCommand.java
@@ -1,16 +1,17 @@
-package command.UserPlaylistCommand;
+package command.playlist.guild;
 
 import YouTubeSearchApi.YoutubeClient;
 import YouTubeSearchApi.entity.YoutubeVideo;
 import com.jagrosh.jdautilities.command.CommandEvent;
-import database.UserPlaylistModel;
-import service.music.CustomEmbedBuilder;
+import database.PlaylistModel;
+import database.PremiumModel;
+import net.dv8tion.jda.api.EmbedBuilder;
 import service.music.HelpProcess;
 
 import java.sql.SQLException;
 import java.util.concurrent.CompletableFuture;
 
-public class AddTrackToPlaylistCommand extends UserPlaylistBaseCommand
+public class AddTrackToPlaylistCommand extends GuildPlaylistBaseCommand
 {
     private final YoutubeClient client;
 
@@ -18,20 +19,34 @@ public class AddTrackToPlaylistCommand extends UserPlaylistBaseCommand
     {
         this.client = ytc;
 
-        this.name = "add_track_to_my_playlist";
-        this.aliases = new String[]{"attmp"};
+        this.name = "add_track_to_guild_playlist";
+        this.aliases = new String[]{"attgp"};
         this.arguments = "<playlist name>, <url>";
-        this.help = "Add a new track to user playlist. \n" +
+        this.help = "Add a new track to guild playlist. \n" +
                     "Use coma (,) as separator for each arguments.";
         this.cooldown = 2;
-        this.category = new Category("User Playlist");
+        this.guildOnly = true;
+        this.category = new Category("Guild Playlist");
         this.help = HelpProcess.getHelp(this);
     }
 
     @Override
     protected void execute(CommandEvent event)
     {
-        CustomEmbedBuilder embed = new CustomEmbedBuilder();
+        EmbedBuilder embed = new EmbedBuilder();
+        embed.setColor(event.getMember().getColor());
+        PremiumModel premium = new PremiumModel();
+
+        if (premium.isPremium(event.getGuild().getIdLong(), this.table) == false)
+        {
+            embed.setTitle("Attention");
+            embed.addField(
+                    ":warning:",
+                    "You are not premium, you can't use this command.",
+                    true);
+            event.reply(embed.build());
+            return;
+        }
 
         if (event.getArgs().split(",").length != 2)
         {
@@ -48,10 +63,10 @@ public class AddTrackToPlaylistCommand extends UserPlaylistBaseCommand
 
         String playlistName = event.getArgs().split(",")[0].trim().replace("'", "\\'");
         String url = event.getArgs().split(",")[1].trim();
+        
+        PlaylistModel db = new PlaylistModel();
 
-        UserPlaylistModel db = new UserPlaylistModel();
-
-        if (db.isPlaylistNameExist(event.getAuthor().getIdLong(), playlistName) == false)
+        if (db.isPlaylistNameAvailable(event.getGuild().getIdLong(), playlistName, this.table))
         {
             embed.setTitle("Attention");
             embed.addField(
@@ -62,15 +77,15 @@ public class AddTrackToPlaylistCommand extends UserPlaylistBaseCommand
             return;
         }
 
-        int userPlaylistTrackCount = db.countPlaylistTrack(
-                db.getPlaylistId(event.getAuthor().getIdLong(), playlistName));
+        int guildPlaylistTrackCount = db.countPlaylistTrack(
+                db.getPlaylistId(event.getGuild().getIdLong(), playlistName, this.table), this.table);
 
-        if (userPlaylistTrackCount >= this.maxTrack)
+        if (guildPlaylistTrackCount >= this.maxTrack)
         {
             embed.setTitle("Failed");
             embed.addField(
                     ":x:",
-                    "Track for playlist has reached the maximum limit.",
+                    "Track for playlist has reached maximum limit.",
                     true);
             event.reply(embed.build());
             return;
@@ -103,14 +118,14 @@ public class AddTrackToPlaylistCommand extends UserPlaylistBaseCommand
             return;
         }
 
-        long playlistId = db.getPlaylistId(event.getAuthor().getIdLong(), playlistName);
+        long playlistId = db.getPlaylistId(event.getGuild().getIdLong(), playlistName, this.table);
         video.setTitle(video.getTitle().replace("'", "\\'"));
 
         CompletableFuture.runAsync(() ->
         {
             try
             {
-                db.addTrackToPlaylist(playlistId, video.getUrl(), video.getTitle());
+                db.addTrackToPlaylist(playlistId, video.getUrl(), video.getTitle(), this.table);
 
                 embed.setTitle("Success");
                 embed.addField(
