+++ src/main/java/command/playlist/guild/RenamePlaylistCommand.java
@@ -1,4 +1,4 @@
-package command.GuildPlaylistCommand;
+package command.playlist.guild;
 
 import com.jagrosh.jdautilities.command.CommandEvent;
 import database.PlaylistModel;
@@ -9,15 +9,15 @@ import service.music.HelpProcess;
 import java.sql.SQLException;
 import java.util.concurrent.CompletableFuture;
 
-public class DeleteTrackFromPlaylistCommand extends GuildPlaylistBaseCommand
+public class RenamePlaylistCommand extends GuildPlaylistBaseCommand
 {
-    public DeleteTrackFromPlaylistCommand()
+    public RenamePlaylistCommand()
     {
-        this.name = "delete_track_from_guild_playlist";
-        this.aliases = new String[]{"dtfgp"};
-        this.arguments = "<playlist name>, <track index>";
-        this.help = "Delete existing track from guild playlist. \n" +
-                    "Use coma (,) as separator for each arguments.";
+        this.name = "rename_guild_playlist";
+        this.aliases = new String[]{"rgp"};
+        this.arguments = "<old playlist name> , <new playlist name>";
+        this.help = "Rename guild playlist. \n" +
+                    "Use coma (,) as separator for old and new playlist name.";
         this.cooldown = 2;
         this.guildOnly = true;
         this.category = new Category("Guild Playlist");
@@ -31,7 +31,7 @@ public class DeleteTrackFromPlaylistCommand extends GuildPlaylistBaseCommand
         embed.setColor(event.getMember().getColor());
         PremiumModel premium = new PremiumModel();
 
-        if (premium.isPremium(event.getGuild().getIdLong(), this.table) == false)
+        if (!premium.isPremium(event.getGuild().getIdLong(), this.table))
         {
             embed.setTitle("Attention");
             embed.addField(
@@ -48,54 +48,38 @@ public class DeleteTrackFromPlaylistCommand extends GuildPlaylistBaseCommand
             embed.addField(
                     ":warning:",
                     "Invalid given arguments.\n" +
-                          "This command need 2 arguments: <playlist name>, <track index>.\n" +
-                          "Use coma (,) as separator for each arguments.",
+                            "This command need 2 arguments: <old playlist name> , <new playlist name>.\n" +
+                            "Use coma (,) as separator for each arguments.",
                     true);
             event.reply(embed.build());
             return;
         }
 
-        String playlistName = event.getArgs().split(",")[0].trim().replace("'", "\\'");
-        int index = Integer.parseInt(event.getArgs().split(",")[1].trim());
+        String oldName = event.getArgs().split(",")[0].trim().replace("'", "\\'");
+        String newName = event.getArgs().split(",")[1].trim().replace("'", "\\'");
 
         PlaylistModel db = new PlaylistModel();
-        int playlistSize = db.countPlaylistTrack(db.getPlaylistId(event.getGuild().getIdLong(), playlistName, this.table), this.table);
 
-        if (playlistSize <= 0)
+        if (db.isPlaylistNameAvailable(event.getGuild().getIdLong(), oldName, this.table))
         {
             embed.setTitle("Failed");
             embed.addField(
                     ":x:",
-                    "`" + playlistName + "` playlist is empty, no track to delete.",
+                    "Playlist `" + oldName + "` not exist.",
                     true);
-            event.reply(embed.build());
             return;
         }
-
-        if (index > (playlistSize))
-        {
-            embed.setTitle("Failed");
-            embed.addField(
-                    ":x:",
-                    "`" + playlistName + "` playlist only have " + playlistSize + " track(s).",
-                    true);
-            event.reply(embed.build());
-            return;
-        }
-
-        long playlistId = db.getPlaylistId(event.getAuthor().getIdLong(), playlistName, this.table);
-        long playlistTrackId = db.getPlaylistTrackId(playlistId, index, this.table);
 
         CompletableFuture.runAsync(() ->
         {
             try
             {
-                db.deleteTrackFromPlaylistAsync(playlistTrackId, this.table);
+                db.renamePlaylist(event.getGuild().getIdLong(), oldName, newName, this.table);
 
                 embed.setTitle("Success");
                 embed.addField(
                         ":white_check_mark:",
-                        "Track deleted from `" + playlistName + "` playlist.",
+                        "Playlist renamed from `" + oldName + "` to `" + newName + "`.",
                         true);
             }
             catch (SQLException e)
@@ -105,7 +89,7 @@ public class DeleteTrackFromPlaylistCommand extends GuildPlaylistBaseCommand
                 embed.setTitle("Failed");
                 embed.addField(
                         ":x:",
-                        "Can't delete track from `" + playlistName + "` playlist.",
+                        "Can't rename playlist.",
                         true);
             }
