+++ src/main/java/command/playlist/user/DeletePlaylistCommand.java
@@ -1,4 +1,4 @@
-package command.UserPlaylistCommand;
+package command.playlist.user;
 
 import com.jagrosh.jdautilities.command.CommandEvent;
 import database.UserPlaylistModel;
@@ -8,15 +8,14 @@ import service.music.HelpProcess;
 import java.sql.SQLException;
 import java.util.concurrent.CompletableFuture;
 
-public class CreatePlaylistCommand extends UserPlaylistBaseCommand
+public class DeletePlaylistCommand extends UserPlaylistBaseCommand
 {
-    public CreatePlaylistCommand()
+    public DeletePlaylistCommand()
     {
-        this.name = "create_new_playlist";
-        this.aliases = new String[]{"cnp"};
+        this.name = "delete_my_playlist";
+        this.aliases = new String[]{"dmp"};
         this.arguments = "<playlist name>";
-        this.help = "Create a new user playlist. \n" +
-                    "Playlist name can't be the same with your other playlist name.";
+        this.help = "Delete the existing user playlist.";
         this.cooldown = 2;
         this.category = new Category("User Playlist");
         this.help = HelpProcess.getHelp(this);
@@ -27,38 +26,26 @@ public class CreatePlaylistCommand extends UserPlaylistBaseCommand
     {
         CustomEmbedBuilder embed = new CustomEmbedBuilder();
 
-        if (event.getArgs().trim().length() == 0)
+        if (event.getArgs().trim().length() <= 0)
         {
             embed.setTitle("Attention");
             embed.addField(
                     ":warning:",
-                    "Please specify a name for the playlist.",
-                    true);
-            event.reply(embed.build());
-            return;
-        }
-
-        UserPlaylistModel db = new UserPlaylistModel();
-
-        if (db.countPlaylist(event.getAuthor().getIdLong()) >= this.maxPlaylist)
-        {
-            embed.setTitle("Failed");
-            embed.addField(
-                    ":x:",
-                    "You have reached the maximum limit for playlist allocated to each user.",
+                    "Please provide the name of the playlist you want to delete.",
                     true);
             event.reply(embed.build());
             return;
         }
 
         String playlistName = event.getArgs().trim().replace("'", "\\'");
+        UserPlaylistModel db = new UserPlaylistModel();
 
-        if (db.isPlaylistNameExist(event.getAuthor().getIdLong(), playlistName))
+        if (db.isPlaylistNameExist(event.getAuthor().getIdLong(), playlistName) == false)
         {
             embed.setTitle("Failed");
             embed.addField(
                     ":x:",
-                    "You already have playlist with the same name.",
+                    "`" + playlistName + "` playlist does not exist.",
                     true);
             event.reply(embed.build());
             return;
@@ -68,12 +55,12 @@ public class CreatePlaylistCommand extends UserPlaylistBaseCommand
         {
             try
             {
-                db.createPlaylist(event.getAuthor().getIdLong(), playlistName);
+                db.deletePlaylist(event.getAuthor().getIdLong(), playlistName);
 
                 embed.setTitle("Success");
                 embed.addField(
                         ":white_check_mark:",
-                        "`" + playlistName + "` playlist is created.",
+                        "`" + playlistName + "` playlist deleted.",
                         true);
             }
             catch (SQLException e)
@@ -83,7 +70,7 @@ public class CreatePlaylistCommand extends UserPlaylistBaseCommand
                 embed.setTitle("Failed");
                 embed.addField(
                         ":x:",
-                        "There's already playlist with name `" + playlistName + "`.",
+                        "`" + playlistName + "` playlist not deleted.",
                         true);
             }
