+++ src/main/java/command/playlist/user/DeleteTrackFromPlaylistCommand.java
@@ -1,4 +1,4 @@
-package command.UserPlaylistCommand;
+package command.playlist.user;
 
 import com.jagrosh.jdautilities.command.CommandEvent;
 import database.UserPlaylistModel;
